+++ src/test/java/org/apache/sysds/test/component/matrix/SeqTableTest.java
@@ -24,7 +24,8 @@ import static org.junit.Assert.assertEquals;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.sysds.runtime.DMLRuntimeException;
-import org.apache.sysds.runtime.matrix.data.LibMatrixTable;
+import org.apache.sysds.runtime.compress.lib.CLALibRexpand;
+import org.apache.sysds.runtime.matrix.data.LibMatrixReorg;
 import org.apache.sysds.runtime.matrix.data.MatrixBlock;
 import org.apache.sysds.test.TestUtils;
 import org.junit.Test;
@@ -34,31 +35,31 @@ public class SeqTableTest {
 	protected static final Log LOG = LogFactory.getLog(SeqTableTest.class.getName());
 
 	static{
-		LibMatrixTable.ALLOW_COMPRESSED_TABLE_SEQ = false; // allow the compressed tables.
+		CLALibRexpand.ALLOW_COMPRESSED_TABLE_SEQ = false; // allow the compressed tables.
 	}
 
 	@Test(expected = DMLRuntimeException.class)
 	public void test_notSameDim() {
 		MatrixBlock c = new MatrixBlock(20, 1, 0.0);
-		LibMatrixTable.tableSeqOperations(10, c, 0);
+		LibMatrixReorg.fusedSeqRexpand(10, c, 0);
 	}
 
 	@Test(expected = DMLRuntimeException.class)
 	public void test_toLow() {
 		MatrixBlock c = new MatrixBlock(10, 1, -1.0);
-		LibMatrixTable.tableSeqOperations(10, c, 0);
+		LibMatrixReorg.fusedSeqRexpand(10, c, 0);
 	}
 
 	@Test(expected = DMLRuntimeException.class)
 	public void test_toManyColumn() {
 		MatrixBlock c = new MatrixBlock(10, 2, -1.0);
-		LibMatrixTable.tableSeqOperations(10, c, 0);
+		LibMatrixReorg.fusedSeqRexpand(10, c, 0);
 	}
 
 	@Test
 	public void test_All_NaN() {
 		MatrixBlock c = new MatrixBlock(10, 1, Double.NaN);
-		MatrixBlock ret = LibMatrixTable.tableSeqOperations(10, c, 1);
+		MatrixBlock ret = LibMatrixReorg.fusedSeqRexpand(10, c, 1);
 
 		assertEquals(0, ret.getNumColumns());
 	}
@@ -66,14 +67,14 @@ public class SeqTableTest {
 	@Test
 	public void test_w_NaN() {
 		MatrixBlock c = new MatrixBlock(10, 1, 1.0);
-		MatrixBlock ret = LibMatrixTable.tableSeqOperations(10, c, Double.NaN);
+		MatrixBlock ret = LibMatrixReorg.fusedSeqRexpand(10, c, Double.NaN);
 		assertEquals(0, ret.getNumColumns());
 	}
 
 	@Test
 	public void test_all_one() {
 		MatrixBlock c = new MatrixBlock(10, 1, 1.0);
-		MatrixBlock ret = LibMatrixTable.tableSeqOperations(10, c, 1);
+		MatrixBlock ret = LibMatrixReorg.fusedSeqRexpand(10, c, 1);
 		assertEquals(1, ret.getNumColumns());
 		TestUtils.compareMatrices(c, ret, 0);
 	}
@@ -81,7 +82,7 @@ public class SeqTableTest {
 	@Test
 	public void test_all_one_givenMatrixBlock() {
 		MatrixBlock c = new MatrixBlock(10, 1, 1.0);
-		MatrixBlock ret = LibMatrixTable.tableSeqOperations(10, c, 1, new MatrixBlock(), true);
+		MatrixBlock ret = LibMatrixReorg.fusedSeqRexpand(10, c, 1, new MatrixBlock(), true);
 		assertEquals(1, ret.getNumColumns());
 		TestUtils.compareMatrices(c, ret, 0);
 	}
@@ -89,7 +90,7 @@ public class SeqTableTest {
 	@Test
 	public void test_all_one_givenMatrixBlockWithSize() {
 		MatrixBlock c = new MatrixBlock(10, 1, 1.0);
-		MatrixBlock ret = LibMatrixTable.tableSeqOperations(10, c, 1, new MatrixBlock(1,2, 0.0), false);
+		MatrixBlock ret = LibMatrixReorg.fusedSeqRexpand(10, c, 1, new MatrixBlock(1,2, 0.0), false);
 		assertEquals(2, ret.getNumColumns());
 		MatrixBlock expected = c.append(new MatrixBlock(10, 1, 0.0));
 		TestUtils.compareMatrices(expected, ret, 0);
@@ -98,7 +99,7 @@ public class SeqTableTest {
 	@Test
 	public void test_all_one_givenMatrixBlockWithSize_NaNWeight() {
 		MatrixBlock c = new MatrixBlock(10, 1, 1.0);
-		MatrixBlock ret = LibMatrixTable.tableSeqOperations(10, c, Double.NaN, new MatrixBlock(1,2, 0.0), false);
+		MatrixBlock ret = LibMatrixReorg.fusedSeqRexpand(10, c, Double.NaN, new MatrixBlock(1,2, 0.0), false);
 		assertEquals(2, ret.getNumColumns());
 		MatrixBlock expected = new MatrixBlock(10, 2, 0.0);
 		TestUtils.compareMatrices(expected, ret, 0);
